Change Features from util into a service spec
This fixes #216
