+++ org.osgi.service.feature/.classpath
