+++ org.osgi.service.feature/.project
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <projectDescription>
-	<name>org.osgi.util.feature</name>
+	<name>org.osgi.service.feature</name>
 	<comment></comment>
 	<projects>
 	</projects>
