+++ org.osgi.service.feature/.settings/org.eclipse.core.resources.prefs
