+++ org.osgi.service.feature/.settings/org.eclipse.core.runtime.prefs
