+++ org.osgi.service.feature/src/org/osgi/service/feature/BuilderFactory.java
@@ -15,7 +15,7 @@
  *
  * SPDX-License-Identifier: Apache-2.0 
  *******************************************************************************/
-package org.osgi.util.feature;
+package org.osgi.service.feature;
 
 import org.osgi.annotation.versioning.ProviderType;
