+++ org.osgi.service.feature/src/org/osgi/service/feature/Feature.java
@@ -15,7 +15,7 @@
  *
  * SPDX-License-Identifier: Apache-2.0 
  *******************************************************************************/
-package org.osgi.util.feature;
+package org.osgi.service.feature;
 
 import java.util.List;
 import java.util.Map;
