+++ org.osgi.service.feature/src/org/osgi/service/feature/FeatureBundle.java
@@ -15,33 +15,21 @@
  *
  * SPDX-License-Identifier: Apache-2.0 
  *******************************************************************************/
-package org.osgi.util.feature;
+package org.osgi.service.feature;
 
 import java.util.Map;
 
 import org.osgi.annotation.versioning.ProviderType;
 
 /**
- * Represents an OSGi Configuration in the Feature Model.
+ * A Bundle which is part of a feature.
  * @ThreadSafe
  */
 @ProviderType
-public interface FeatureConfiguration {
+public interface FeatureBundle extends FeatureArtifact {
     /**
-     * Get the PID from the configuration.
-     * @return The PID.
+     * Get the metadata for this bundle.
+     * @return The metadata.
      */
-    String getPid();
-
-    /**
-     * Get the Factory PID from the configuration, if any.
-     * @return The Factory PID, or {@code null} if there is none.
-     */
-    String getFactoryPid();
-
-    /**
-     * Get the configuration key-value map.
-     * @return The key-value map.
-     */
-    Map<String, Object> getValues();
+    Map<String, Object> getMetadata();
 }
