+++ org.osgi.service.feature/src/org/osgi/service/feature/FeatureConfiguration.java
@@ -15,21 +15,33 @@
  *
  * SPDX-License-Identifier: Apache-2.0 
  *******************************************************************************/
-package org.osgi.util.feature;
+package org.osgi.service.feature;
 
 import java.util.Map;
 
 import org.osgi.annotation.versioning.ProviderType;
 
 /**
- * A Bundle which is part of a feature.
+ * Represents an OSGi Configuration in the Feature Model.
  * @ThreadSafe
  */
 @ProviderType
-public interface FeatureBundle extends FeatureArtifact {
+public interface FeatureConfiguration {
     /**
-     * Get the metadata for this bundle.
-     * @return The metadata.
+     * Get the PID from the configuration.
+     * @return The PID.
      */
-    Map<String, Object> getMetadata();
+    String getPid();
+
+    /**
+     * Get the Factory PID from the configuration, if any.
+     * @return The Factory PID, or {@code null} if there is none.
+     */
+    String getFactoryPid();
+
+    /**
+     * Get the configuration key-value map.
+     * @return The key-value map.
+     */
+    Map<String, Object> getValues();
 }
